Node: Implement KAD store RPC. - Routes and stores to K nodes, if possible. - Storage backend still missing (simple HashMap ATM).

+++ src/routing/constants.rs
@@ -16,4 +16,4 @@ pub const MAX_ERRORS: usize = 5;
 pub const MAX_RTT_MS: u64 = 1500;
 
 /// Maximum concurrency per operation
-pub const KAD_ALPHA: i32 = 3;
+pub const KAD_ALPHA: usize = 3;
